Phone Number
Start of work on Phone Number VO, including validation for UK phone
numbers. Not sure about the methods in the validation objects at the
moment, need to consider whether this is the best place for them or not.
We're currently using the validation object to identify variations, such
as UK and US phone numbers, so since this will have the logic to
understand whether the format is correct or not, it also makes sense for
the validator to be able to sanitise and to provide versions of the
number, such as the international dialing version, for example